Resolve circular dependency for ToolEvaluator
by placing ComputeEnvironment and ActionBox under job_execution.
