+++ lib/galaxy/tools/evaluation.py
@@ -4,15 +4,15 @@ import os
 import shlex
 import string
 import tempfile
-from typing import Any, Dict, List
+from datetime import datetime
+from typing import Any, Callable, Dict, List, Optional
 
 from galaxy import model
 from galaxy.files import ProvidesUserFileSourcesUserContext
+from galaxy.job_execution.compute_environment import ComputeEnvironment
 from galaxy.job_execution.setup import ensure_configs_directory
-from galaxy.jobs import ComputeEnvironment
 from galaxy.model.none_like import NoneDataset
 from galaxy.security.object_wrapper import wrap_with_safe_string
-from galaxy.tools import global_tool_errors
 from galaxy.tools.parameters import (
     visit_input_values,
     wrapped_json,
@@ -90,6 +90,25 @@ What won't work:
 - Can't just serialize param_dict and overwrite inputs (I think), too many custom methods that could be called
 
 """
+
+
+class ToolErrorLog:
+    def __init__(self):
+        self.error_stack = []
+        self.max_errors = 100
+
+    def add_error(self, file, phase, exception):
+        self.error_stack.insert(0, {
+            "file": file,
+            "time": str(datetime.now()),
+            "phase": phase,
+            "error": unicodify(exception)
+        })
+        if len(self.error_stack) > self.max_errors:
+            self.error_stack.pop()
+
+
+global_tool_errors = ToolErrorLog()
 
 
 class ToolEvaluator:
