+++ setup.cfg
@@ -203,8 +203,6 @@ check_untyped_defs = False
 check_untyped_defs = False
 [mypy-galaxy.model.custom_types]
 check_untyped_defs = False
-[mypy-galaxy.jobs.actions.post]
-check_untyped_defs = False
 [mypy-galaxy.job_metrics.collectl.processes]
 check_untyped_defs = False
 [mypy-galaxy.datatypes.util.gff_util]
